also switch retry_deferred
